refactor: use url search params, better naming
